+++ lib/gorilla/render_target.h
@@ -21,7 +21,7 @@ public:
   bool create_rtv(const ComPtr<ID3D11Device> &device,
                   const ComPtr<ID3D11Texture2D> &texture);
   bool create_dsv(const ComPtr<ID3D11Device> &device);
-  void setup(const ComPtr<ID3D11DeviceContext> &context, int w, int h);
+  void setup(const ComPtr<ID3D11DeviceContext> &context, float w, float h);
 };
 
 } // namespace gorilla
